test: activate application prior to launch
tetst: add basic checks for foreground execution and hittability

When running UI tests we expect the application to be running in the
foreground in order to be able to interact with its on-screen elements

Add a check to make sure the application runs in foreground.

Add checks to make sure elements are on-screen and can actually be
interacted with.

test: play video only after view acually appears on screen

test: add application launch wait check

test: add on-screen element existence check